Add files via upload
The program implements the program Hanoi Towers with Graphical User Interface in asm language and intel syntax, it was developed back in 2007 as the final project of a University subject in National Autonomous University of Mexico (UNAM). It was compiled with Borland Turbo Assembler (TASM). The executable runs on Windows 7 and lower.

The program includes a timer and a move counter so that you can compare your performance with other places, you can select from 3 to 9 disks on the Hanoi Towers.